+++ discord/cogs/sostickets/__init__.py
@@ -2,4 +2,4 @@ from .sostickets import SosTickets
 
 
 async def setup(bot):
-    await bot.add_cog(SosTickets())
+    await bot.add_cog(SosTickets(bot))